refactor: 피드백1, 예외처리 메시지 출력에 대한 추상화 레벨 동일하게 수정
