+++ src/main/kotlin/domain/lotto/LottoApplication.kt
@@ -8,13 +8,15 @@ import domain.lotto.service.LottoService
 import domain.lotto.strategy.LottoRandomShuffleStrategy
 import domain.lotto.ui.LottoInputView
 import domain.lotto.ui.LottoResultView
+import global.common.ui.ExceptionView
 import global.strategy.split.CommaSplitStrategy
 import global.strategy.ui.ConsoleInputStrategy
 import global.strategy.ui.ConsoleOutputStrategy
 
 class LottoApplication(
     private val lottoInputView: LottoInputView,
-    private val lottoResultView: LottoResultView
+    private val lottoResultView: LottoResultView,
+    private val exceptionView: ExceptionView
 ) {
     fun run() {
         val money = purchaseLottoByConsole()
@@ -32,7 +34,7 @@ class LottoApplication(
         return try {
             WinningLotto.from(lottoByConsole(), bonusBallByConsole())
         } catch (e: Exception) {
-            ConsoleOutputStrategy.output(e.message.toString())
+            exceptionView.output(e.message.toString())
             winningLottoByConsole()
         }
     }
@@ -41,7 +43,7 @@ class LottoApplication(
         return try {
             Money(lottoInputView.purchaseLotto())
         } catch (e: Exception) {
-            ConsoleOutputStrategy.output(e.message.toString())
+            exceptionView.output(e.message.toString())
             purchaseLottoByConsole()
         }
     }
@@ -50,7 +52,7 @@ class LottoApplication(
         return try {
             Lotto.of(lottoInputView.winningLotto(), CommaSplitStrategy)
         } catch (e: Exception) {
-            ConsoleOutputStrategy.output(e.message.toString())
+            exceptionView.output(e.message.toString())
             lottoByConsole()
         }
     }
@@ -59,7 +61,7 @@ class LottoApplication(
         return try {
             LottoNumber.of(lottoInputView.bonusBall())
         } catch (e: Exception) {
-            ConsoleOutputStrategy.output(e.message.toString())
+            exceptionView.output(e.message.toString())
             bonusBallByConsole()
         }
     }
@@ -68,6 +70,7 @@ class LottoApplication(
 fun main() {
     val lottoInputView = LottoInputView(ConsoleInputStrategy, ConsoleOutputStrategy)
     val lottoResultView = LottoResultView(ConsoleOutputStrategy)
-    val lottoApplication = LottoApplication(lottoInputView, lottoResultView)
+    val exceptionView = ExceptionView(ConsoleOutputStrategy)
+    val lottoApplication = LottoApplication(lottoInputView, lottoResultView, exceptionView)
     lottoApplication.run()
 }
