+++ src/main/kotlin/global/common/ui/ExceptionView.kt
@@ -1,0 +1,7 @@
+package global.common.ui
+
+import global.strategy.ui.OutputStrategy
+
+class ExceptionView(private val outputStrategy: OutputStrategy) {
+    fun output(message: String) = outputStrategy.output(message)
+}
